+++ app/Http/Livewire/Admin/Roles/Edit.php
@@ -6,9 +6,12 @@ use App\Models\Admin\Permissions\Role;
 use Livewire\Component;
 use Livewire\WithPagination;
 use App\Models\Admin\Log;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Edit extends Component
 {
+    use AuthorizesRequests;
+
     public Role $role;
     public $readyToLoad = false;
     public $search;
@@ -24,6 +27,8 @@ class Edit extends Component
 
     public function RoleForm()
     {
+        $this->authorize('role-edit',Role::class);
+
         $this->validate();
         $this->role->update($this->validate());
         $this->role->permissions()->sync($this->permissions);
@@ -35,6 +40,8 @@ class Edit extends Component
 
     public function render()
     {
+        $this->authorize('roles',Role::class);
+
         $role = $this->role;
         $roles = $this->readyToLoad ? Role::where('title', 'LIKE', '%' . $this->search . '%')->orWhere('description', 'LIKE', '%' . $this->search . '%')->latest()->paginate(5) : [];
 
@@ -53,6 +60,8 @@ class Edit extends Component
 
     public function delete()
     {
+        $this->authorize('role-delete',Role::class);
+
         $role = Role::find($this->deleteId);
         $role->delete();
