+++ resources/views/livewire/admin/permissions/edit.blade.php
@@ -53,16 +53,18 @@
                         <div class="card">
                             <div class="card-body">
                                 <h4 class="card-title mb-2">لیست سطوح دسترسی ها</h4>
-                                <a href="{{ route('admin.permissions.trash') }}" type="button"
-                                    class="btn btn-danger mb-2 mr-2" style="float:left;margin-top:-37px;"><i
-                                        class="fa fa-refresh"></i> سطل زباله
-                                    <span class="badge badge-danger">
-                                        {{ \App\Models\Admin\Permissions\Permission::onlyTrashed()->count() }}
-                                    </span>
-                                </a>
-                                <button type="button" class="btn btn-primary mb-2 mr-2"
+                                @can('permission-trash')
+                                    <a href="{{ route('admin.permissions.trash') }}" type="button"
+                                        class="btn btn-danger mb-2 mr-2" style="float:left;margin-top:-37px;"><i
+                                            class="fa fa-refresh"></i> سطل زباله
+                                        <span class="badge badge-danger">
+                                            {{ \App\Models\Admin\Permissions\Permission::onlyTrashed()->count() }}
+                                        </span>
+                                    </a>
+                                @endcan
+                                {{-- <button type="button" class="btn btn-primary mb-2 mr-2"
                                     style="float:left;margin-top:-37px;"><i class="fa fa-file-excel-o"></i> خروجی
-                                    اکسل</button>
+                                    اکسل</button> --}}
                                 <a href="{{ route('admin.permissions') }}" class="btn btn-success mb-2 mr-2"
                                     style="float:left;margin-top:-37px;"><i class="fa fa-plus-square"></i> افزودن</a>
 
@@ -94,13 +96,17 @@
                                                         @endforeach
                                                     </td>
                                                     <td>
-                                                        <a href="{{ route('admin.permissions.edit', $permission->id) }}"
-                                                            class="action-icon"> <i
-                                                                class="zmdi zmdi-edit zmdi-custom"></i></a>
-                                                        <button wire:click="deleteId({{ $permission->id }})"
-                                                            data-toggle="modal" data-target="#exampleModal"
-                                                            class="action-icon"> <i
-                                                                class="zmdi zmdi-delete zmdi-custom"></i></button>
+                                                        @can('permission-edit')
+                                                            <a href="{{ route('admin.permissions.edit', $permission->id) }}"
+                                                                class="action-icon"> <i
+                                                                    class="zmdi zmdi-edit zmdi-custom"></i></a>
+                                                        @endcan
+                                                        @can('permission-delete')
+                                                            <button wire:click="deleteId({{ $permission->id }})"
+                                                                data-toggle="modal" data-target="#exampleModal"
+                                                                class="action-icon"> <i
+                                                                    class="zmdi zmdi-delete zmdi-custom"></i></button>
+                                                        @endcan
                                                     </td>
                                                 </tr>
                                             @endforeach
